WEBWMS-15 Event Subscriber für Last Login hinzugefügt und Vorbereitung für ACL

+++ public/assets/js/webWmsCoreJs/jquery.user.js
@@ -66,32 +66,34 @@
         ]
     });
 
-    $.contextMenu({
-        selector: 'tr',
-        trigger: 'right',
-        callback: function(key, options, event) {
-            const row = userTable.row(options.$trigger);
+    if (user_role === 'ROLE_SUPER_ADMIN') {
+        $.contextMenu({
+            selector: 'tr',
+            trigger: 'right',
+            callback: function(key, options, event) {
+                const row = userTable.row(options.$trigger);
 
-            switch (key) {
-                case 'edit' :
-                    editUser(row.data().username);
-                    break;
-                case 'editPassword' :
-                    editUserPassword(row.data().username);
-                    break;
-                case 'delete' :
-                    deleteUser(row.data().username);
-                    break;
-                default :
-                    break
+                switch (key) {
+                    case 'edit' :
+                        editUser(row.data().username);
+                        break;
+                    case 'editPassword' :
+                        editUserPassword(row.data().username);
+                        break;
+                    case 'delete' :
+                        deleteUser(row.data().username);
+                        break;
+                    default :
+                        break
+                }
+            },
+            items: {
+                'edit': {name: 'Bearbeiten', icon: 'edit'},
+                'editPassword': {name: 'Passwort ändern', icon: 'edit'},
+                'delete': {name: 'Löschen', icon: 'delete'}
             }
-        },
-        items: {
-            'edit': {name: 'Bearbeiten', icon: 'edit'},
-            'editPassword': {name: 'Passwort ändern', icon: 'edit'},
-            'delete': {name: 'Löschen', icon: 'delete'}
-        }
-    });
+        });
+    }
 
     $(function(){
         // Ändern der Standardbreite des Modals
